% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hagfish_function.R
\name{hagfish_function}
\alias{hagfish_function}
\title{A hagfish function}
\usage{
hagfish_function(love = T)
}
\arguments{
\item{love}{Do you love hagfish? Defaults to TRUE.}
}
\description{
This function allows you to express your love of hagfish.
}
\examples{
hagfish_function()
}
\keyword{hagfish}
